Add one-click deploy script and update tests
