+++ DIFF_2025-06-24_03-09-07.md
@@ -1,0 +1,73 @@
+diff --git a/README.md b/README.md
+index 53ad80c..9e25af8 100644
+--- a/README.md
++++ b/README.md
+@@ -39,6 +39,12 @@ docker compose up -d --build
+ # api:  http://localhost:4000
+ ```
+ 
++### One-click deploy
++Use the helper script to run tests and start all services in Docker:
++```bash
++./scripts/one_click_deploy.sh
++```
++
+ ## Deployment
+ 
+ CI/CD is handled by GitHub Actions (see `.github/workflows/ci.yml`). On pushes to `main`, the workflow builds the frontend, builds a Docker image and pushes it to GHCR, then SSH deploys to a Dell R720 server running Docker Compose.
+diff --git a/frontend/src/App.test.tsx b/frontend/src/App.test.tsx
+index bf4b739..fb045ec 100644
+--- a/frontend/src/App.test.tsx
++++ b/frontend/src/App.test.tsx
+@@ -1,16 +1,11 @@
+ import { render, screen } from '@testing-library/react';
+ import React from 'react';
+ import { describe, it, expect } from 'vitest';
+-import { MemoryRouter } from 'react-router-dom';
+ import App from './App';
+ 
+ describe('App', () => {
+   it('renders Dumbo component', () => {
+-    render(
+-      <MemoryRouter>
+-        <App />
+-      </MemoryRouter>
+-    );
++    render(<App />);
+     expect(screen.getByText(/Meet\s+Dumbo!/i)).toBeTruthy();
+   });
+ });
+diff --git a/scripts/one_click_deploy.sh b/scripts/one_click_deploy.sh
+new file mode 100755
+index 0000000..9162999
+--- /dev/null
++++ b/scripts/one_click_deploy.sh
+@@ -0,0 +1,28 @@
++#!/bin/bash
++set -e
++
++# Load environment variables from .env if present
++if [ -f .env ]; then
++  export $(grep -v '^#' .env | xargs)
++fi
++
++# Ensure required environment variables
++missing=false
++for var in CF_API_TOKEN CLOUDFLARED_TOKEN; do
++  if [ -z "${!var}" ]; then
++    echo "Missing required env var: $var"
++    missing=true
++  fi
++done
++if [ "$missing" = true ]; then
++  exit 1
++fi
++
++# Run tests
++npm test --prefix backend
++npm test --prefix frontend
++
++# Build and launch containers
++docker compose up -d --build
++
++echo "Deployment complete."
