+++ README.md
@@ -39,6 +39,12 @@ docker compose up -d --build
 # api:  http://localhost:4000
 ```
 
+### One-click deploy
+Use the helper script to run tests and start all services in Docker:
+```bash
+./scripts/one_click_deploy.sh
+```
+
 ## Deployment
 
 CI/CD is handled by GitHub Actions (see `.github/workflows/ci.yml`). On pushes to `main`, the workflow builds the frontend, builds a Docker image and pushes it to GHCR, then SSH deploys to a Dell R720 server running Docker Compose.
