+++ RECOMMENDATIONS_2025-06-24_03-09-07.md
@@ -1,0 +1,3 @@
+- Add .env.example documenting CF_API_TOKEN and CLOUDFLARED_TOKEN for easier setup
+- Consider extending one_click_deploy.sh to run database migrations automatically
+- Add more frontend tests to cover routing
