+++ frontend/index.html
@@ -7,6 +7,6 @@
   </head>
   <body class="bg-gray-900 text-white">
     <div id="root"></div>
-    <script type="module" src="/src/main.jsx"></script>
+    <script type="module" src="/src/main.tsx"></script>
   </body>
 </html>
