+++ frontend/package.json
@@ -11,7 +11,8 @@
   "dependencies": {
     "framer-motion": "^11.0.0",
     "react": "^18.3.0",
-    "react-dom": "^18.3.0"
+    "react-dom": "^18.3.0",
+    "react-router-dom": "^6.23.0"
   },
   "devDependencies": {
     "@testing-library/react": "^14.1.2",
