+++ frontend/src/App.tsx
@@ -1,0 +1,28 @@
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
+import { motion } from 'framer-motion';
+import Navbar from './components/Navbar';
+import Home from './pages/Home';
+import Gallery from './pages/Gallery';
+import About from './pages/About';
+
+export default function App() {
+  return (
+    <Router>
+      <div className="min-h-screen bg-gray-900 text-white">
+        <Navbar />
+        <motion.main
+          initial={{ opacity: 0 }}
+          animate={{ opacity: 1 }}
+          transition={{ duration: 1 }}
+          className="p-8"
+        >
+          <Routes>
+            <Route path="/" element={<Home />} />
+            <Route path="/gallery" element={<Gallery />} />
+            <Route path="/about" element={<About />} />
+          </Routes>
+        </motion.main>
+      </div>
+    </Router>
+  );
+}
