+++ frontend/src/components/Navbar.tsx
@@ -1,24 +1,26 @@
+import { Link } from 'react-router-dom';
+
 export default function Navbar() {
   return (
     <nav className="bg-black py-4 px-6 flex justify-between items-center shadow-md">
       <div className="text-neon-pink font-extrabold text-2xl">
-        iongiveafuq.com
+        <Link to="/">iongiveafuq.com</Link>
       </div>
       <ul className="flex space-x-4">
         <li>
-          <a href="#" className="hover:text-neon-green">
+          <Link to="/" className="hover:text-neon-green">
             Home
-          </a>
+          </Link>
         </li>
         <li>
-          <a href="#" className="hover:text-neon-green">
-            Store
-          </a>
+          <Link to="/gallery" className="hover:text-neon-green">
+            Gallery
+          </Link>
         </li>
         <li>
-          <a href="#" className="hover:text-neon-green">
+          <Link to="/about" className="hover:text-neon-green">
             About
-          </a>
+          </Link>
         </li>
       </ul>
     </nav>
