+++ frontend/src/components/mascots/Dumbo.tsx
