+++ frontend/src/pages/About.tsx
@@ -1,0 +1,8 @@
+export default function About() {
+  return (
+    <div className="p-8 text-center">
+      <h2 className="text-3xl font-bold text-neon-blue mb-4">About Us</h2>
+      <p>iongiveafuq.com is a dark-humor experiment and art gallery.</p>
+    </div>
+  );
+}
