+++ frontend/src/pages/Gallery.tsx
@@ -1,0 +1,9 @@
+import ArtworkViewer from '../components/viewer/ArtworkViewer';
+
+export default function Gallery() {
+  return (
+    <div className="p-4">
+      <ArtworkViewer />
+    </div>
+  );
+}
