+++ infra/caddy/Caddyfile
@@ -1,0 +1,8 @@
+iongiveafuq.com {
+    root * /srv
+    encode gzip
+    file_server
+    handle_path /api/* {
+        reverse_proxy http://api:4000
+    }
+}
