+++ infra/nginx/nginx.conf
@@ -1,0 +1,25 @@
+user nginx;
+events {}
+http {
+    upstream app {
+        server web:80;
+        server caddy:80;
+    }
+
+    server {
+        listen 80;
+        server_name iongiveafuq.com;
+
+        location /api/ {
+            proxy_pass http://api:4000;
+            proxy_set_header Host $host;
+            proxy_set_header X-Real-IP $remote_addr;
+        }
+
+        location / {
+            proxy_pass http://app;
+            proxy_set_header Host $host;
+            proxy_set_header X-Real-IP $remote_addr;
+        }
+    }
+}
